requirements/base.txt
Django==5.1.6
djangorestframework==3.15.2
psycopg2-binary==2.9.10
python-decouple==3.8
gunicorn==23.0.0
